+++ CHANGES.md
@@ -13,6 +13,7 @@
  * Improved availability of code suggestions in certain situations.
  * Improved accuracy of code suggestions in predicate arguments.
  * Errors, warnings, and infos logged while loading JSON & SNBT schemas are now logged using their respective log level.
+ * Deleting a file using the Files tab now also closes the associated document.
 
 ### Fixes
  * Fixed drifting of encoded index when parsing nested JSON with a nesting depth > 2.